Smoothened the movement of the enemies

+++ src/com/zainic/zainiship/path/Path.java
@@ -2,20 +2,20 @@ package com.zainic.zainiship.path;
 
 public class Path {
 	
-	public static int pathOneLeftX(int t) {
-		return (int) (0.05*t + 1);
+	public static double pathOneLeftX(double t) {
+		return (0.05*t + 1);
 	}
 	
-	public static int pathOneLeftY(int t) {
-		return (int) (5 - 0.05*t);
+	public static double pathOneLeftY(double t) {
+		return (5 - 0.05*t);
 	}
 	
-	public static int pathOneRightX(int t) {
-		return (int) (-0.05*t - 1);
+	public static double pathOneRightX(double t) {
+		return (-0.05*t - 1);
 	}
 	
-	public static int pathOneRightY(int t) {
-		return (int) (5 - 0.05*t);
+	public static double pathOneRightY(double t) {
+		return (5 - 0.05*t);
 	}
 	
 }
